requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.2
django-analytical==2.5.0
django-anymail==6.0.1
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-email-obfuscator==0.1.5
gunicorn==19.9.0
idna==2.8
psycopg2-binary==2.8.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
